refactor: presentational 컴포넌트에 global 폴더 추가
appbar, header, footer등이 global로 들어가고 나머지는 main 폴더로 이동
