+++ user-web/src/components/presentational/global/footer/LicenseTextBox.jsx
