+++ user-web/src/components/presentational/main/banner/HomeBanner.jsx
